+++ main code/main/Due/Coms_Serial.h
@@ -59,7 +59,7 @@ class Coms_Serial: public Coms {
     void init_serial();
 
     void write_frame(int address);
-    void send_menu_frame(int menu, int encoder_pos);
+    void send_menu_frame(byte menu, int encoder_pos);
     void send_all_text_frames(bool send_now = false);
     void send_all_sensor_data_frames(bool send_now = false);
     void send_text_frame(byte obj_num, byte address = 0);
